♻️ move driver types to their own header and reduce redundancy in exports
