+++ include/qdmi/client/control.h
@@ -15,6 +15,7 @@ SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #include "qdmi/common/types.h"
 #include "qdmi/device/types.h"
+#include "qdmi/driver/types.h"
 
 #ifdef __cplusplus
 #include <cstddef>
