+++ include/qdmi/driver/session.h
@@ -15,7 +15,7 @@ SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #pragma once
 
-#include "qdmi/common/types.h"
+#include "qdmi/driver/types.h"
 
 #ifdef __cplusplus
 #include <cstddef>
