+++ test/utils/test_impl.hpp
@@ -15,7 +15,7 @@ SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #pragma once
 
-#include "qdmi/common/types.h"
+#include "qdmi/driver/types.h"
 
 #include <gtest/gtest.h>
 #include <string>
